renamed BaseApi to K8sApi and refactored methods
